fix + test: fix main test for verification through sdk
This is a cheap solution by adding a day to the verification timestamp. I removed the precondition check that checks whether the current time is greater than the time at which the stamp was last verified, but this should not have any significant impact on the validity of the GetExpiration function.